Move subscriber for data container
