+++ src/DataContainer/PaletteController.php
@@ -12,7 +12,7 @@
  * @copyright Copyright 2016 ContaoBlackForest
  */
 
-namespace ContaoBlackForest\Contao\Core\DcGeneral\Controller;
+namespace ContaoBlackForest\Contao\Core\DcGeneral\DataContainer;
 
 use Contao\System;
 use ContaoBlackForest\Contao\Core\DcGeneral\Service\TableToGeneralService;
@@ -34,7 +34,7 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 /**
  * Class PaletteController
  *
- * @package ContaoBlackForest\Contao\Core\DcGeneral\Palette
+ * @package ContaoBlackForest\Contao\Core\DcGeneral\DataContainer
  */
 class PaletteController implements EventSubscriberInterface
 {
